feat: build Luckfox Pico code via CMake

+++ luckfox_pico/include/features.hpp
@@ -6,26 +6,26 @@
 #include <cmath>
 #include <cctype>
 
-inline double digit_ratio(const std::string &s)
+inline float digit_ratio(const std::string &s)
 {
     if (s.empty())
         return 0.0;
     size_t count = 0;
     for (unsigned char ch : s) if (std::isdigit(ch)) ++count;
-    return static_cast<double>(count) / s.size();
+    return static_cast<float>(count) / s.size();
 }
 
-inline double shannon_entropy(const std::string &s)
+inline float shannon_entropy(const std::string &s)
 {
     if (s.empty())
         return 0.0;
     std::unordered_map<unsigned char, size_t> freq;
     for (unsigned char ch : s) ++freq[ch];
-    const double n = static_cast<double>(s.size());
-    double ent = 0.0;
+    const float n = static_cast<float>(s.size());
+    float ent = 0.0;
     for (const auto &p : freq)
     {
-        double f = static_cast<double>(p.second) / n;
+        float f = static_cast<float>(p.second) / n;
         ent -= f * std::log2(f);
     }
     return ent;
@@ -36,12 +36,12 @@ inline size_t length(const std::string &s)
     return s.size();
 }
 
-inline double uppercase_ratio(const std::string &s)
+inline float uppercase_ratio(const std::string &s)
 {
     if (s.empty())
         return 0.0;
     size_t count = 0;
     for (unsigned char ch : s) if (std::isupper(ch)) ++count;
-    return static_cast<double>(count) / s.size();
+    return static_cast<float>(count) / s.size();
 }
 #endif
